Default formatting config for JS/TS (#118)
These are the current settings I have for the built-in
javascript/typescript formatter for vscode. We can of course tweak or
consider another extension as needed.

I included the formatting settings that are true but not the ones that
are default false. If that becomes a problem we can add the false ones
in too.

Closes #117