Replace buttons with MUI-based interface (#143)
- Created buy button and buy modal to replace the previous BuyMenu
- Sell, Undo, Redo and Start buttons also replaced with MUI buttons

For convenience, the "B" key can be used to open the Buy Modal. "Esc"
can be used to cancel any pending purchase. I plan to cleanup the layout
and formatting in a future PR, after completing some of the other
migrations.

I'm thinking we should eventually break out Fill and Dig into their own
buttons as well.



![image](https://github.com/user-attachments/assets/d0c5e378-d17e-43c3-a08c-2eafe2d6d01c)

Part of #142